# public/docs/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Lumen: Loja de vendas",
        "version": "0.1"
    },
    "paths": {
...
        "/sales_item/create": {
            "post": {
                "tags": [
                    "Itens de venda"
                ],
                "summary": "Cadastrar novo item de venda",
                "operationId": "87e97ea23af50e8572634dd986235566",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "sales_item_date",
                                    "sale_id",
                                    "product_id",
                                    "quantity"
                                ],
                                "properties": {
                                    "sales_item_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-04-06"
                                    },
                                    "sale_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "product_id": {
                                        "type": "integer",
                                        "example": "2"
                                    },
                                    "": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Item de venda criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "sales_item_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "sales_item_date": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2025-04-06"
                                        },
                                        "sale_id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "product_id": {
                                            "type": "integer",
                                            "example": "2"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Dados inválidos"
                    }
                }
            }
        },
        "/sales_item/{sale_id}": {
            "get": {
                "tags": [
                    "Itens de venda"
                ],
                "summary": "Buscar itens de venda de uma venda",
                "operationId": "182bab18b16fa8eecffb71c6b8fc613d",
                "parameters": [
                    {
                        "name": "sale_id",
                        "in": "path",
                        "description": "Id da venda",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Aqui estão os itens desta venda",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "sales_item_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "sales_item_date": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "2025-04-06"
                                            },
                                            "sale_id": {
                                                "type": "integer",
                                                "example": "1"
                                            },
                                            "product_id": {
                                                "type": "integer",
                                                "example": "2"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Dados inválidos"
                    }
                }
            }
        },
        "/sales_item/update/{id}": {
            "put": {
                "tags": [
                    "Itens de venda"
                ],
                "summary": "Alterar um cadastro",
                "operationId": "a26a302b314c5972ec7277280f964271",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do item de venda",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "sales_item_date",
                                    "sale_id",
                                    "product_id",
                                    "quantity"
                                ],
                                "properties": {
                                    "sales_item_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-04-06"
                                    },
                                    "sale_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "product_id": {
                                        "type": "integer",
                                        "example": "2"
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Item de venda alterado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "sales_item_date": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2025-04-06"
                                        },
                                        "sale_id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "product_id": {
                                            "type": "integer",
                                            "example": "2"
                                        },
                                        "quantity": {
                                            "type": "integer",
                                            "example": "1"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Dados inválidos"
                    }
                }
            }
        },
        "/sales_item/delete/{sales_item_id}": {
            "delete": {
                "tags": [
                    "Itens de venda"
                ],
                "summary": "Deletar um item de venda pelo ID",
                "operationId": "54d7e58556a767ea270781c246bcaaeb",
                "parameters": [
                    {
                        "name": "sales_item_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Item de venda deletado com sucesso"
                    },
                    "404": {
                        "description": "Item de venda não encontrado"
                    }
                }
            }
        }
    },
...
}